bug 553433 - update freetype to version 2.3.12 in our tree. r=blassey
UltraBlame original commit: 228a9faee7e828e27145ec9245ede3467564b54b

+++ modules/freetype2/src/tools/apinames.c
@@ -265,7 +265,7 @@ read_header_file( FILE*  file, int  verbose )
           if ( p > name )
           {
             if ( verbose )
-              fprintf( stderr, ">>> %.*s\n", p-name, name );
+              fprintf( stderr, ">>> %.*s\n", (int)(p - name), name );
 
             names_add( name, p );
           }
